Archer!

+++ Crater.java
@@ -7,7 +7,7 @@ public class Crater extends Sprite{
 	int damage = 0;  //Variable for the amount of damage the sprite does
 	int lives = 1;  //Variable for the amount of lives the sprite has
 
-	public Crater(int x, int y, int xl, int yl, String filepath) {
+	public Crater(double x, double y, int xl, int yl, String filepath) {
 		super(filepath, x, y, xl, yl);
 	}
