+++ Lake.java
@@ -8,7 +8,7 @@ public class Lake extends Sprite{
 	int damage = 0;  //Variable for the amount of damage the sprite does
 	int lives = 2;  //Variable for the amount of lives the sprite has
 
-	public Lake(int x, int y, int xl, int yl, String filepath) {
+	public Lake(double x, double y, int xl, int yl, String filepath) {
 		super(filepath, x, y, xl, yl);
 	}
